+++ src/main/java/com/scb/bank/app/service/FundTransferService.java
@@ -2,5 +2,5 @@ package com.scb.bank.app.service;
 
 public interface FundTransferService {
 
-    String doFundTransfer(long fromAcc, long toAccNo, int amountToTransfer);
+    String doFundTransfer(long fromAcc, long toAccNo, float amountToTransfer);
 }
